examples/python-template/content/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
{% if values.databaseType == 'postgres' -%}
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.35
{% elif values.databaseType == 'mysql' -%}
pymysql>=1.1.1
sqlalchemy>=2.0.35
{% elif values.databaseType == 'mongodb' -%}
pymongo>=4.10.1
motor>=3.6.0
{% elif values.databaseType == 'redis' -%}
redis>=5.2.0
{% endif -%}
